Added extensions folder and improved guessing game.
