+++ main.py
@@ -1,6 +1,7 @@
 import discord
 from discord.ext import commands
 import yaml
+import os
 
 with open('config.yml', 'r') as file:
   config = yaml.safe_load(file)
@@ -17,8 +18,9 @@ def get_prefix(bot, message):
 bot = commands.Bot(command_prefix='.', description='CodiHacks bot')
 
 # Load cogs by the filenames
-bot.load_extension('guess')
-bot.load_extension('rng')
+for i in os.listdir('extensions'):
+	if i.endswith('.py') and os.path.isfile('extensions/' + i):
+		bot.load_extension('extensions.' + i.replace('.py', ''))
 
 # Event listeners
 @bot.event
